backend/pyproject.toml
[tool.poetry]
name = "collage-generator-backend"
version = "1.0.0"
description = "Collage Generator Application"
authors = ["Sergey Dyachok <sergey@sdyachok.com.ua>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13.2"
uvicorn = "^0.11.3"
aiofiles = "^0.4.0"
requests = "^2.23.0"
python-multipart = "^0.0.5"
asyncpg = "^0.20.1"
databases = "^0.2.6"
sqlalchemy = "^1.3.15"
python-dotenv = "^0.12.0"
alembic = "^1.4.2"
pytz = "^2019.3"
psycopg2 = "^2.8.4"
marshmallow = "^3.5.1"
starlette-context = "^0.2.0"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
pyjwt = "^1.7.1"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
pytest-cov = "^2.8.1"
pytest-xdist = "^1.31.0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.eggs
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | docs
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
